Work (#48)
* map

* map

* map

* searchMap

* map

* map

* map

* map

* map

* map

* test

* map

* map

* map

* mapJsInit

* address parse to location

* remove a dependency

* restaurant into map

* main to mapview

* work

* modify overlay

* getDatas

* getDataList

* tmp

* createRestaurantDataAroundMe

* location rebuild

* interprogress

* mark around me

---------

+++ src/main/java/com/project/team/Restaurant/RestaurantService.java
@@ -72,14 +72,14 @@ public class RestaurantService {
         return restaurantRepository.findByAddress(address);
     }
 
-    public List<Restaurant> getAround(String x, String y, Long radius) {
-        Long myX = Long.parseLong(x);
-        Long myY = Long.parseLong(y);
+    public List<Restaurant> getAround(String x, String y, double radius) {
+        double myX = Double.parseDouble(x);
+        double myY = Double.parseDouble(y);
         List<Restaurant> restaurantList = restaurantRepository.findAll();
         List<Restaurant> restaurants = new ArrayList<>();
         for (Restaurant restaurant : restaurantList) {
-            Long resX = Long.parseLong(restaurant.getLocationX());
-            Long resY = Long.parseLong(restaurant.getLocationY());
+            double resX = Double.parseDouble(restaurant.getLocationX());
+            double resY = Double.parseDouble(restaurant.getLocationY());
             if (resX < myX + radius && resX > myX - radius && resY < myY + radius && resY > myY - radius)
                 restaurants.add(restaurant);
         }
